Anagram_Java
A simple anagram algorithm, for explanatory purposes. 
Algorithm calculates the number of an individual letter in each string.
Then compares the counts in both strings, logging the lower count. This allows removes excess totals that might exist in one string.
Two times the final count is then subtracted from the the total length of both strings.